bitcoin-core-sv2: fix u32 overflow in coinbase_weight
coinbase_output_max_additional_size is a peer-controlled u32 from
CoinbaseOutputConstraints, and WEIGHT_FACTOR was also u32, so the
multiplication was evaluated in u32 and only widened to u64 afterwards.
Values >= 0x4000_0001 wrapped silently in release builds (0x4000_0000
wrapped to 0 and clamped to MIN_BLOCK_RESERVED_WEIGHT) and panicked in
debug builds.

Widen WEIGHT_FACTOR to u64 and cast before multiplying, in both the
v30x and v31x trees.

Closes #676

+++ bitcoin-core-sv2/src/unix_capnp/mod.rs
@@ -14,7 +14,7 @@ pub mod v31x;
 const MIN_BLOCK_RESERVED_WEIGHT: u64 = 2000;
 
 /// BIP141 weight factor (witness scale factor), used to convert vsize to weight units.
-const WEIGHT_FACTOR: u32 = 4;
+const WEIGHT_FACTOR: u64 = 4;
 
 /// Grace period before stale template data is retired after a chain tip change, in seconds.
 ///
